requirements.txt
PyYAML
dnslib